[module6-task3]: Move data-access services to /data-access/services
